medicare_analysis/requirements.txt
pandas==2.1.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.27.0
python-dotenv==1.0.0

